# docs/openapi.json
{
  "components": {
    "schemas": {
...
      "ErrorEnvelope": {
        "properties": {
          "error": {
            "properties": {
              "codigo": {
                "description": "Código estable (ver HANDOFF_FRONTEND.md §7)",
                "type": "string"
              },
              "detalles": {
                "description": "Objeto, lista o null; nunca incluye el input recibido"
              },
              "mensaje": {
                "description": "Texto humano; puede cambiar",
                "type": "string"
              },
              "request_id": {
                "description": "También en el header X-Request-ID",
                "type": "string"
              }
            },
            "required": [
              "codigo",
              "mensaje",
              "detalles",
              "request_id"
            ],
            "type": "object"
          }
        },
        "required": [
          "error"
        ],
        "title": "ErrorEnvelope",
        "type": "object"
      },
...
    },
    "securitySchemes": {
      "HTTPBearer": {
        "scheme": "bearer",
        "type": "http"
      },
      "bearerAuth": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "description": "Contrato HTTP del backend. Prefijo /v1. Ver docs/HANDOFF_FRONTEND.md.",
    "title": "Módulo 1 — Documentación habilitante",
    "version": "1.0.0-rc1",
    "x-migracion-head": "0016_plantillas_globales"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/v1/consultas/backlog_oc": {
      "get": {
        "operationId": "backlog_oc_v1_consultas_backlog_oc_get",
        "parameters": [
          {
            "in": "query",
            "name": "estado",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "default": "activo",
              "title": "Estado"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Backlog Oc V1 Consultas Backlog Oc Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Backlog Oc"
      }
    },
    "/v1/consultas/cobertura_oc": {
      "get": {
        "operationId": "cobertura_oc_v1_consultas_cobertura_oc_get",
        "parameters": [
          {
            "in": "query",
            "name": "commitment_id",
            "required": true,
            "schema": {
              "title": "Commitment Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Cobertura Oc V1 Consultas Cobertura Oc Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Cobertura Oc"
      }
    },
    "/v1/consultas/decision": {
      "get": {
        "operationId": "decision_v1_consultas_decision_get",
        "parameters": [
          {
            "in": "query",
            "name": "referencia_evaluacion",
            "required": true,
            "schema": {
              "title": "Referencia Evaluacion",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decision V1 Consultas Decision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Decision"
      }
    },
    "/v1/consultas/decisiones_oc": {
      "get": {
        "operationId": "decisiones_oc_v1_consultas_decisiones_oc_get",
        "parameters": [
          {
            "in": "query",
            "name": "commitment_id",
            "required": true,
            "schema": {
              "title": "Commitment Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Decisiones Oc V1 Consultas Decisiones Oc Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Decisiones Oc"
      }
    },
    "/v1/consultas/historial_supervision": {
      "get": {
        "operationId": "historial_supervision_v1_consultas_historial_supervision_get",
        "parameters": [
          {
            "in": "query",
            "name": "sujeto_id",
            "required": true,
            "schema": {
              "title": "Sujeto Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Historial Supervision V1 Consultas Historial Supervision Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Historial Supervision"
      }
    },
    "/v1/consultas/incumplimiento_empresa": {
      "get": {
        "operationId": "incumplimiento_empresa_v1_consultas_incumplimiento_empresa_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Incumplimiento Empresa V1 Consultas Incumplimiento Empresa Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Incumplimiento Empresa"
      }
    },
    "/v1/consultas/legajo": {
      "get": {
        "operationId": "legajo_v1_consultas_legajo_get",
        "parameters": [
          {
            "in": "query",
            "name": "sujeto_id",
            "required": true,
            "schema": {
              "title": "Sujeto Id",
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Legajo V1 Consultas Legajo Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Legajo"
      }
    },
    "/v1/consultas/log_auditoria": {
      "get": {
        "operationId": "log_auditoria_v1_consultas_log_auditoria_get",
        "parameters": [
          {
            "in": "query",
            "name": "tipo",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Tipo"
            }
          },
          {
            "in": "query",
            "name": "desde",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Desde"
            }
          },
          {
            "in": "query",
            "name": "hasta",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "format": "date-time",
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Hasta"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 500,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Log Auditoria V1 Consultas Log Auditoria Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Log Auditoria"
      }
    },
...
    "/v1/consultas/plantillas_globales": {
      "get": {
        "description": "Plantillas de industria junto a las copias locales (estado sin_copia / al_dia /\nactualizacion_disponible) para decidir a mano qué traer (no-funcionales 1.5.3).",
        "operationId": "plantillas_globales_v1_consultas_plantillas_globales_get",
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "additionalProperties": true,
                  "title": "Response Plantillas Globales V1 Consultas Plantillas Globales Get",
                  "type": "object"
                }
              }
            },
            "description": "Successful Response"
          },
          "401": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "No autenticado (sin token, vencido, usuario inexistente o inactivo)"
          },
          "403": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Rol insuficiente o fuera de alcance"
          },
          "404": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Recurso inexistente o no visible para el rol"
          },
          "409": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Conflicto de dominio, idempotencia o concurrencia"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Validación o regla de dominio"
          },
          "500": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ErrorEnvelope"
                }
              }
            },
            "description": "Error interno; informar request_id"
          }
        },
        "security": [
          {
            "bearerAuth": []
          }
        ],
        "summary": "Plantillas Globales"
      }
    },
...
  }
}